+++ headers/nu_xml_parser.h
@@ -447,9 +447,9 @@ static void NU_Apply_Node_Defaults(struct Node* node)
     node->gap = 0.0f;
     node->content_width = 0.0f;
     node->content_height = 0.0f;
-    node->index = UINT32_MAX;
-    node->parent_index = UINT32_MAX;
-    node->first_child_index = UINT32_MAX;
+    node->index = UINT16_MAX;
+    node->parent_index = UINT16_MAX;
+    node->first_child_index = UINT16_MAX;
     node->child_capacity = 0;
     node->child_count = 0;
     node->node_present = 1;
